ssi endpoints
